refactor(bootstrap): keep the skill package check behind the host facade
The package fixture added with this check builds both host manifests, so
the names .codex-plugin and .claude-plugin entered main.rs and the guard
for INV.PLATFORM.OS-BEHIND-FACADE failed: host knowledge belongs behind
crates/unica-bootstrap/src/host/, and that rule has no path exemptions.

Move verify_installed_skill_package into host::skill_package next to
verify_installed_plugin_metadata, which it already calls and whose family
of package verification it belongs to. Its tests move with it, as the
invariant requires of an adapter's tests. The binary now passes the
release version in and holds no host knowledge of its own.
